+++ src/types/store.d.ts
@@ -1,0 +1,7 @@
+declare module 'storeTypes' {
+  interface sliceState {
+    auth: authState;
+    user: userState;
+    toggle: toggleState;
+  }
+}
